requirements.txt
fastapi
pydantic
python-dotenv
google-generativeai